+++ iterator/src/blk_metadata.rs
@@ -4,12 +4,12 @@ use crate::path_to_modified_time;
 
 #[derive(Debug, Clone, Copy)]
 pub struct BlkMetadata {
-    pub index: usize,
+    pub index: u16,
     pub modified_time: u64,
 }
 
 impl BlkMetadata {
-    pub fn new(index: usize, path: &Path) -> Self {
+    pub fn new(index: u16, path: &Path) -> Self {
         Self {
             index,
             modified_time: path_to_modified_time(path),
